python/productivity_tracker/requirements.txt
Django==3.2.19
pytz==2021.1
jsonpickle==2.0.0
google-cloud-language==2.2.2
google-api-python-client==2.15.0
google-auth==1.34.0
gunicorn==20.1.0
pymysql==1.0.2
django-environ==0.4.5
google-cloud-secret-manager==2.7.0
